# multi: update main package and btcutil to chainhash/v1.1.0, use optimized dsha256 