@Roasbeef 2023-12-19
Here's a benchmark run vs master: 
```
benchmark                         old ns/op     new ns/op     delta
BenchmarkCalcSigHash-8            1081499       1244944       +15.11%
BenchmarkCalcWitnessSigHash-8     34499         32224         -6.59%

benchmark                         old allocs     new allocs     delta
BenchmarkCalcSigHash-8            712            623            -12.50%
BenchmarkCalcWitnessSigHash-8     445            801            +80.00%

benchmark                         old bytes     new bytes     delta
BenchmarkCalcSigHash-8            1290508       933085        -27.70%
BenchmarkCalcWitnessSigHash-8     46992         19936         -57.58%
```