@Roasbeef 2023-12-19
Here's another comparison, old is w/o the last commit, new is w/ it: 
```
benchmark                          old ns/op     new ns/op     delta
BenchmarkCalcSigHash-32            1866235       1975372       +5.85%
BenchmarkCalcWitnessSigHash-32     65135         45449         -30.22%

benchmark                          old allocs     new allocs     delta
BenchmarkCalcSigHash-32            623            623            +0.00%
BenchmarkCalcWitnessSigHash-32     534            801            +50.00%

benchmark                          old bytes     new bytes     delta
BenchmarkCalcSigHash-32            933080        933081        +0.00%
BenchmarkCalcWitnessSigHash-32     58384         19936         -65.85%
```