@Roasbeef 2023-12-19
Here's a run with the same computer as the first run above, but this time with `GOAMD64=v3`: 
```
benchmark                          old ns/op     new ns/op     delta
BenchmarkCalcSigHash-32            1989442       1837011       -7.66%
BenchmarkCalcWitnessSigHash-32     59216         43734         -26.14%

benchmark                          old allocs     new allocs     delta
BenchmarkCalcSigHash-32            712            623            -12.50%
BenchmarkCalcWitnessSigHash-32     445            801            +80.00%

benchmark                          old bytes     new bytes     delta
BenchmarkCalcSigHash-32            1290506       933081        -27.70%
BenchmarkCalcWitnessSigHash-32     46992         19936         -57.58%
```