@Roasbeef 2023-12-19
The final comparison: 
```
goos: linux
goarch: amd64
pkg: github.com/btcsuite/btcd/txscript
cpu: 13th Gen Intel(R) Core(TM) i9-13900K
                      │   old.txt    │               new.txt               │
                      │    sec/op    │   sec/op     vs base                │
CalcSigHash-32          2.151m ±  9%   1.910m ± 4%  -11.23% (p=0.009 n=10)
CalcWitnessSigHash-32   63.55µ ± 15%   42.25µ ± 6%  -33.53% (p=0.000 n=10)
geomean                 369.8µ         284.0µ       -23.18%

                      │    old.txt    │               new.txt                │
                      │     B/op      │     B/op      vs base                │
CalcSigHash-32          1260.3Ki ± 0%   911.2Ki ± 0%  -27.70% (p=0.000 n=10)
CalcWitnessSigHash-32    45.89Ki ± 0%   18.08Ki ± 0%  -60.61% (p=0.000 n=10)
geomean                  240.5Ki        128.3Ki       -46.63%

                      │  old.txt   │               new.txt                │
                      │ allocs/op  │ allocs/op   vs base                  │
CalcSigHash-32          712.0 ± 0%   623.0 ± 0%  -12.50% (p=0.000 n=10)
CalcWitnessSigHash-32   445.0 ± 0%   445.0 ± 0%        ~ (p=1.000 n=10) ¹
geomean                 562.9        526.5        -6.46%
```